Fix for Case 9 : Compatibility with HTML5 autofocus attribute
The autofocus attribute was causing the form element to be selected
without a focus() event. This was preventing Uni-Form from highlighting
the element. This has been fixed and tested.  Please note that as a side
effect of the testing, we had to add a small hack to the form.js test
file. The autofocus attribute does cause a change in the form value.
This means there is a bug when you use autofocus + ask_on_leave +
default data together. I feel that this is a small enough issue to
ignore for the moment.